git-svn-id: http://svn.ratbox.org/svnroot/trunk@18204 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/m_list.c
@@ -156,7 +156,7 @@ list_limit_channels(struct Client *source_p, const char *param)
 	char *args;
 	char *p;
 	dlink_node *ptr;
-	int sendq_limit;
+	unsigned int sendq_limit;
 	int max = INT_MAX;
 	int min = 0;
 	int i;
@@ -206,8 +206,8 @@ list_limit_channels(struct Client *source_p, const char *param)
 			break;
 		}
 
-		if((int)dlink_list_length(&chptr->members) >= max ||
-		   (int)dlink_list_length(&chptr->members) <= min)
+		if(dlink_list_length(&chptr->members) >= max ||
+		   dlink_list_length(&chptr->members) <= min)
 			continue;
 
 		if(SecretChannel(chptr) && !IsMember(source_p, chptr))
